route account into server url - refactor frontend

+++ spring-boot-accounting-book/src/main/java/com/xintongthecoder/accountingbook/controller/SpendingItemController.java
@@ -23,10 +23,13 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import com.xintongthecoder.accountingbook.dao.AccountBookRepository;
+import com.xintongthecoder.accountingbook.dao.AccountRepository;
 import com.xintongthecoder.accountingbook.dao.SpendingItemRepository;
+import com.xintongthecoder.accountingbook.entity.Account;
 import com.xintongthecoder.accountingbook.entity.AccountBook;
 import com.xintongthecoder.accountingbook.entity.Category;
 import com.xintongthecoder.accountingbook.entity.SpendingItem;
+import com.xintongthecoder.accountingbook.errorHandler.AccountAccessDeniedException;
 import com.xintongthecoder.accountingbook.errorHandler.AccountBookNotFoundException;
 import com.xintongthecoder.accountingbook.errorHandler.SpendingItemNotFoundException;
 import com.xintongthecoder.accountingbook.modelAssembler.SpendingItemModelAssembler;
@@ -37,43 +40,61 @@ import org.springframework.web.bind.annotation.RequestBody;
 
 
 @RestController
-@RequestMapping("api")
+@RequestMapping("api/accounts")
 public class SpendingItemController {
 
+    private final AccountRepository accountRepository;
     private final AccountBookRepository accountBookRepository;
     private final SpendingItemRepository spendingItemRepository;
     private final SpendingItemModelAssembler spendingItemModelAssembler;
     private final PagedResourcesAssembler<SpendingItem> pagedResourcesAssembler;
 
 
-    public SpendingItemController(SpendingItemRepository spendingItemRepository,
+    public SpendingItemController(AccountRepository accountRepository,
             AccountBookRepository accountBookRepository,
+            SpendingItemRepository spendingItemRepository,
             SpendingItemModelAssembler spendingItemModelAssembler,
             PagedResourcesAssembler<SpendingItem> pagedResourcesAssembler) {
+        this.accountRepository = accountRepository;
         this.accountBookRepository = accountBookRepository;
         this.spendingItemRepository = spendingItemRepository;
         this.spendingItemModelAssembler = spendingItemModelAssembler;
         this.pagedResourcesAssembler = pagedResourcesAssembler;
     }
 
-    @GetMapping(value = "/books/{bookId}/items/{itemId}", produces = {"application/hal+json"})
+    @GetMapping(value = "/{email}/books/{bookId}/items/{itemId}",
+            produces = {"application/hal+json"})
     public ResponseEntity<PagedModel<EntityModel<SpendingItem>>> one(
-            @PathVariable(value = "bookId") Long bookId,
-            @PathVariable(value = "itemId") Long itemId,
+            @PathVariable("email") String email, @PathVariable("bookId") Long bookId,
+            @PathVariable("itemId") Long itemId,
             @RequestParam(value = "page", defaultValue = "0") Integer page,
             @RequestParam(value = "size", defaultValue = "10") Integer size) {
+        if (!requestFromAuthorizedAccount(email, bookId)) {
+            throw new AccountAccessDeniedException("item");
+        }
         Page<SpendingItem> pagedItem =
                 spendingItemRepository.findById(itemId, PageRequest.of(page, size));
         return ResponseEntity.ok().contentType(MediaTypes.HAL_JSON)
                 .body(pagedResourcesAssembler.toModel(pagedItem, spendingItemModelAssembler));
     }
 
-    @GetMapping(value = "/books/{bookId}/items", produces = {"application/hal+json"})
-    public ResponseEntity<PagedModel<EntityModel<SpendingItem>>> all(@PathVariable Long bookId,
+    private boolean requestFromAuthorizedAccount(String email, Long bookId) {
+        if (accountBookRepository.getReferenceById(bookId) == null) {
+            throw new AccountBookNotFoundException(bookId);
+        }
+        return accountBookRepository.findById(bookId).get().getAccount().getEmail().equals(email);
+    }
+
+    @GetMapping(value = "/{email}/books/{bookId}/items", produces = {"application/hal+json"})
+    public ResponseEntity<PagedModel<EntityModel<SpendingItem>>> all(
+            @PathVariable("email") String email, @PathVariable("bookId") Long bookId,
             @RequestParam(value = "page", defaultValue = "0") Integer page,
             @RequestParam(value = "size", defaultValue = "10") Integer size,
             @RequestParam(value = "category", required = false) String category,
             @RequestParam(value = "text", required = false) String text) {
+        if (!requestFromAuthorizedAccount(email, bookId)) {
+            throw new AccountAccessDeniedException("items");
+        }
         Page<SpendingItem> pagedItems = spendingItemRepository
                 .findAll(getFilters(bookId, category, text), PageRequest.of(page, size));
         return ResponseEntity.ok().contentType(MediaTypes.HAL_JSON)
@@ -104,31 +125,38 @@ public class SpendingItemController {
         };
     }
 
-    @PostMapping(value = "/books/{bookId}/items")
-    public ResponseEntity<SpendingItem> addItem(@PathVariable Long bookId,
-            @RequestBody SpendingItem itemFromRequest) {
-        Optional<AccountBook> book = this.accountBookRepository.findById(bookId);
-        if (book.isEmpty()) {
-            throw new AccountBookNotFoundException(bookId);
+    @PostMapping(value = "{email}/books/{bookId}/items")
+    public ResponseEntity<SpendingItem> addItem(@PathVariable("email") String email,
+            @PathVariable Long bookId, @RequestBody SpendingItem itemFromRequest) {
+        if (!requestFromAuthorizedAccount(email, bookId)) {
+            throw new AccountAccessDeniedException("items");
         }
-        itemFromRequest.setBook(book.get());
+        AccountBook book = accountBookRepository.findById(bookId).get();
+        itemFromRequest.setBook(book);
         SpendingItem newItem = this.spendingItemRepository.save(itemFromRequest);
         return new ResponseEntity<>(newItem, HttpStatus.CREATED);
     }
 
-    @PutMapping(value = "/items/{itemId}")
-    public ResponseEntity<SpendingItem> editItem(@PathVariable("itemId") Long itemId,
-            @RequestBody SpendingItem itemFromRequest) {
-        if (spendingItemRepository.getReferenceById(itemId) == null) {
-            throw new SpendingItemNotFoundException(itemId);
+    @PutMapping(value = "{email}/items/{itemId}")
+    public ResponseEntity<SpendingItem> editItem(@PathVariable("email") String email,
+            @PathVariable("itemId") Long itemId, @RequestBody SpendingItem itemFromRequest) {
+        Long bookId = spendingItemRepository.getReferenceById(itemId).getBook().getId();
+        if (!requestFromAuthorizedAccount(email, bookId)) {
+            throw new AccountAccessDeniedException("item");
         }
-        itemFromRequest.setBook(spendingItemRepository.getReferenceById(itemId).getBook());
+        AccountBook book = accountBookRepository.getReferenceById(bookId);
+        itemFromRequest.setBook(book);
         SpendingItem updatedItem = spendingItemRepository.save(itemFromRequest);
         return new ResponseEntity<>(updatedItem, HttpStatus.OK);
     }
 
-    @DeleteMapping(value = "/items/{itemId}")
-    public ResponseEntity<HttpStatus> deleteItem(@PathVariable(value = "itemId") Long itemId) {
+    @DeleteMapping(value = "{email}/items/{itemId}")
+    public ResponseEntity<HttpStatus> deleteItem(@PathVariable("email") String email,
+            @PathVariable("itemId") Long itemId) {
+        Long bookId = spendingItemRepository.getReferenceById(itemId).getBook().getId();
+        if (!requestFromAuthorizedAccount(email, bookId)) {
+            throw new AccountAccessDeniedException("item");
+        }
         spendingItemRepository.deleteById(itemId);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
